Move settings into it's own class as we will want to expand it for future work
